fix next build and split receivables client
